Cleaned up breakfast route and renamed Recipes.js to recipe.js in model folder
